added Dark Mode capability for all marketing emails
Finished adding Dark Mode capability (simulated for web browsers using JavaScript) for all marketing emails. Added class tags to all BG and Copy elements so that if you choose to do so, you can toggle between light and dark mode with a button when viewing in a web browser - or force dark mode "on load". Dark mode is simulated because in email, the email client itself changes the code depending on user settings - so the emails were not coded to natively switch when viewed in a web browser since email does not support JavaScript.